// src/generated/models/environment-model.ts
import type { Application } from './application-model';

export const EnvironmentEnvironmentTypeKeyToLabel = {
  'Development': 'Development',
  'Test': 'Test',
  'Production': 'Production'
} as const;
export type EnvironmentEnvironmentTypeKey = keyof typeof EnvironmentEnvironmentTypeKeyToLabel;

export interface Environment {
  /**
   * @displayName Environment
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Environment Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  environmentName: string;
  /**
   * @displayName Active State
   * @validationRule Required for create/update operations
   */
  activeState: boolean;
  /**
   * @displayName Application
   * @validationRule Required for create/update operations
   */
  application: Pick<Application, 'id' | 'applicationName'>;
  /**
   * @displayName Environment Code
   * @validationRule Required for create/update operations
   */
  environmentCode: string;
  /**
   * @displayName Environment Type
   * @validationRule Required for create/update operations
   */
  environmentTypeKey: EnvironmentEnvironmentTypeKey;
}

export const _Environment = 'Environment' as const;